Finish `pyproject.toml` migration (#382)
* Gather system `import`s and separate from external

* Reorder external `import`s

* Run `import setuptools`

* Use `setuptools` instead of `distutils`

* Drop explicit `cythonize` calls

As `setuptools` will handle this for us, go ahead and drop the
`cythonize` calls.

* Use Cython's `build_ext` (if available)

* Drop `setup_requires`

Since these build requirements are already in `pyproject.toml` and that
is the preferred way forward, drop them from `setup.py`.

* Move setuptool-scm config to `pyproject.toml`

* Move nearly all of `setup.py` to `pyproject.toml`

* Add more URLs

* Add `.txt` extension to license

* Use latest setuptools build backend

* Drop extra whitespace

* Add release note

* Fix lint

* Move `pytest` config options to `pyproject.toml`

* Move `.coveragerc` into `pyproject.toml`

* Standardize on `setuptools-scm` with `-`

* Add `setuptools` & `setuptools-scm` to dev reqs

* Consolidate requirements

* Move `setuptools` after `project` items

* Simplify classifiers

* Drop doc build as it already happens on RTD

This preceeded docs builds on RTD for PRs. Now that RTD does PR builds
for docs, this is redundant. So drop it.

* Move `msgpack` later and unwrap

* Run `flake8` before tests

* Bake coverage & doctest flags into pytest config

* Drop `coveralls` from dependencies

* Fix Codecov release note reference

* Use `int16` in doctests for consistent reprs

Windows defaults to `int32` so won't display those `dtype`s. Whereas
UNIX defaults to `int64` and so won't display those `dtype`s. This
creates mismatches between the representations in docstrings between the
two platforms. For simplicity just use `int16`, which both platforms
will represent the same. This should ensure consistency with doctests.

* Add coverage report output type to config too

* Push `numcodecs` test path into config

* Exclude directories outside of numcodecs

* Simplify `flake8` instructions

* Delete unused `build.cmd`

This was used to configure compilers on Windows in the past.
Particularly it was used to configure Python 2.7 Windows builds. However
as of Python 3.5 and UCRT on Windows, this is unnecessary. Further
Python 2.7 was improved by a one off Windows Python compiler installer.
In any event this can safely be dropped. It is not used in this repo and
shouldn't be needed outside of it.

* Drop upper bound on Python

* Drop `requirements_dev.txt` too

* Always use Cython's `new_build_ext`

This should always be available as `cython` is a build requirement. So
simplify to just use Cython and skip `setuptools` here.

+++ numcodecs/delta.py
@@ -28,14 +28,14 @@ class Delta(Codec):
     --------
     >>> import numcodecs
     >>> import numpy as np
-    >>> x = np.arange(100, 120, 2, dtype='i8')
-    >>> codec = numcodecs.Delta(dtype='i8', astype='i1')
+    >>> x = np.arange(100, 120, 2, dtype='i2')
+    >>> codec = numcodecs.Delta(dtype='i2', astype='i1')
     >>> y = codec.encode(x)
     >>> y
     array([100,   2,   2,   2,   2,   2,   2,   2,   2,   2], dtype=int8)
     >>> z = codec.decode(y)
     >>> z
-    array([100, 102, 104, 106, 108, 110, 112, 114, 116, 118])
+    array([100, 102, 104, 106, 108, 110, 112, 114, 116, 118], dtype=int16)
 
     """
